Rename plants viewModel
