+++ app/src/main/java/com/example/wateringreminder/Modules.kt
@@ -5,6 +5,6 @@ import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
-    viewModel { PlantsViewModel(get()) }
+    viewModel { WateringViewModel(get()) }
     viewModel { PlantCreationViewModel(get()) }
 }
